abstraction 😊
